Address PR review: Merge VectorCube support into _filter.py
Per reviewer feedback (@clausmichele):
- Merged filter_bbox.py functionality into existing _filter.py
- Removed duplicate filter_bbox.py file
- Removed demo scripts (demo_openeo_integration.py, test_netherlands_demo.py, test_vectorcube_example.py)
- Kept Netherlands GeoJSON test data file

Changes to _filter.py:
- Added VectorCube support to filter_bbox function signature (Union[VectorCube, RasterCube])
- Added VectorCube filtering logic with intersects() method (includes boundary geometries)
- Fixed critical coordinate transformation bug in _reproject_bbox:
  * Corrected bbox point order from [lat, lon] to [lon, lat]
  * Fixed transformer.transform call to use (x, y) with always_xy=True
- Maintained full backward compatibility for RasterCube filtering

All existing tests pass (5/5 in test_filter.py)
