--- tests/demo_openeo_integration.py
@@ -1,90 +1,0 @@
-#!/usr/bin/env python3
-"""
-OpenEO Process-Style demonstration of enhanced filter_bbox
-
-This example demonstrates the enhanced filter_bbox in the context of
-OpenEO process graphs and datacube operations, showing how it integrates
-into typical OpenEO workflows.
-"""
-
-import geopandas as gpd
-from openeo_pg_parser_networkx.pg_schema import BoundingBox
-
-from openeo_processes_dask.process_implementations.cubes.filter_bbox import filter_bbox
-
-
-def demonstrate_openeo_process_integration():
-    """
-    Demonstrate how filter_bbox fits into OpenEO process workflows.
-
-    In OpenEO, processes are typically chained like:
-    load_collection() -> filter_temporal() -> filter_bbox() -> apply() -> save_result()
-    """
-    print("🔗 OpenEO Process Integration Example")
-    print("=" * 50)
-
-    # Simulate an OpenEO process graph step:
-    # Step 1: load_collection("vector-boundaries")
-    print("📡 Process: load_collection('administrative-boundaries')")
-
-    # Load real vector data (this simulates an OpenEO vector collection)
-    try:
-        vector_cube = gpd.read_file("Netherlands_polygon.geojson")
-        print(f"   ✅ Loaded VectorCube: {len(vector_cube)} features")
-    except FileNotFoundError:
-        print("   ⚠️  Netherlands_polygon.geojson not found, creating mock data")
-        # Create minimal mock vector data
-        from shapely.geometry import Point
-
-        vector_cube = gpd.GeoDataFrame(
-            {
-                "id": [1, 2],
-                "name": ["Region_A", "Region_B"],
-                "geometry": [Point(5.0, 52.0), Point(6.0, 53.0)],
-            },
-            crs="EPSG:4326",
-        )
-
-    # Step 2: filter_bbox() - This is our enhanced process
-    print("🗺️  Process: filter_bbox(west=4.0, south=51.0, east=6.0, north=53.0)")
-
-    # Define bounding box using OpenEO schema
-    bbox = BoundingBox(west=4.0, south=51.0, east=6.0, north=53.0, crs="EPSG:4326")
-
-    # Apply the enhanced filter_bbox process
-    filtered_cube = filter_bbox(vector_cube, bbox)
-
-    print(f"   ✅ Filtered VectorCube: {len(filtered_cube)} features")
-    print("   📊 Original features:", len(vector_cube))
-    print("   📉 Filtered features:", len(filtered_cube))
-
-    return True
-
-
-def main():
-    """Main demonstration function."""
-    print("🌍 Enhanced filter_bbox: OpenEO Process Integration")
-    print("=" * 60)
-    print()
-    print("This demonstrates how the enhanced filter_bbox integrates")
-    print("into OpenEO process graphs and maintains full compatibility")
-    print("with the OpenEO datacube and process paradigms.")
-    print()
-
-    # Run demonstrations
-    process_success = demonstrate_openeo_process_integration()
-
-    print(f"\n📊 Summary:")
-    print(
-        f"OpenEO Process Integration: {'✅ SUCCESS' if process_success else '❌ FAILED'}"
-    )
-
-    if process_success:
-        print("\n🎯 Enhanced filter_bbox is ready for OpenEO backends!")
-        print("   ✅ Maintains OpenEO API compatibility")
-        print("   ✅ Supports both RasterCube and VectorCube")
-        print("   ✅ Integrates seamlessly into process graphs")
-
-
-if __name__ == "__main__":
-    main()
